+++ packages/espressocash_app/lib/features/ramp/services/off_ramp_order_service.dart
@@ -134,9 +134,7 @@ class OffRampOrderService implements Disposable {
 
           return CryptoAmount(
             value: amount,
-            cryptoCurrency: CryptoCurrency(
-              token: token,
-            ),
+            cryptoCurrency: CryptoCurrency(token: token),
           );
         },
       );