+++ packages/espressocash_app/lib/features/stellar/service/stellar_client.dart
@@ -183,6 +183,10 @@ class StellarClient {
 
     final response = await _sdk.submitTransaction(transaction);
 
+    if (!response.success) {
+      throw Exception('Send Stellar USDC failed: $response');
+    }
+
     return response.success;
   }
 }